newsportal/requirements.txt
Django
psycopg2
django-allauth
django-filter
celery[redis]
redis
python-dotenv
django-modeltranslation